Include a Makefile for the library and the its tests
The libant directory now includes a Makefile for the library itself and
its tests. For now this Makefile includes targets for building the
library itself and for building the tests. Additionally there is also a
target for cleanup. The compiler is set to clang and should't be changed
to ensure that llvm works.

This Makefile is able to build any c file located in the
tests directory. This is done to make adding tests easier and quicker.
This helps the cause of the project because more test allow the library
to be more reliable and up-to-date with its own features.
A tests suite is also a powerful tool for finding bugs before they get
merged and this is a good step towards a test suite.