Libant: Created the buffer struct + some refactoring
This patch modifies the buffer to be a struct instead of just being a
char*. Additionally to the pointer the struct also includes an integer
len which describes the length of the buffer (amount of chars).

The struct is called libant_buffer and is defined in libant/buffer.h.
From this point onwards the buffer will always be passed with the
libant_buffer type and usually by reference.

Additionally this patch updates the existing commands and tests to match
the ways of the new buffer struct.