Libant: Makefile: We now compile the library into a dynamic library
Libant gets now compiled (and installed) into a dynamic library located
in /usr/lib.

This patch adds the necessary targets for installation and
uninstallation and makes the necessary changes to the existing targets
so that the library is compiled into a dynamic library.

Additionally this patch adds the missing .PHONY target
to the makefile so that make doesn't get confised with files that have
the same name as the targets.