Libant: Implemented new coordinate system stored in libant_buffer
This patch implements a new coordinate system thats main purpose is to
keep track of the position in the buffer. This system is implemented via
3 new variables (integers) stored in the libant_buffer struct. These 3
variables are cur_pos, xcoord and ycoord. The x and y coords represent
the 2d position in a file while the cur_pos is equivalent to the
buffer-index that the 2d coords describe.

This patch includes a new helper method that can convert the different
types of coordinates to eachother. This method is called
libant_buffer_update_coordinates() and takes as an argument a
libant_buffer by reference. Thereafter it checks what type of conversion
the caller wants (this is indicated by setting either the 1d coordinate
or the 2d coordinates to -1) and deals with it accordingly.

Lastly this patch adds a new test to test out the new method. This test
is called update-coords and it features a basic implementation of the
newly added method.