Added config loading and parsing functionality
Added multiple methods to handle config loading from config files.
The configuration language is similar to vimscript although much more
simplistic. At this point it only includes 2 methods: set and get.

The configuration is stored in the libant_buffer as an array of
libant_setting structs. libant_setting is defined in config.h together
with a static constant default_settings array which includes all the
setting pared with their default values. On buffer initialization (in
libant_init_buffer()) buffer.settings will be initialized with the
default settings.

Loading a config file, is done with the libant_load_config_file()
method. As an argument it takes the filepath and a libant_buffer by reference.
libant_load_config_file() uses libant_parse_line() internally to walk
through the given config file line by line and set the settings in the
passed libant_buffer according to what is defined in the config file.

libant_parse_line() can additionally be used to parse commands passed by
the user on runtime. If the command return something (like for example
if it is a "get" command). The return value will be written to the
"response" member of the passed libant_buffer.

The available verbs and their internally used command counterparts are
defined in the command.h file and are stored in an array that specifies
the verb used for the command together with the command-function and the
amount of arguments it takes.

Additionally this patch includes a couple of test to test out the
parsing.