Allow string literals with spaces in config
The config system is now able to parse arguments that have spaces in
them. While it is still possible to give string-arguments without the
use of quotation marks the q-mark method is the only way to specify
arguments that include spaces. There will be no intention to demand
users to use q-marks for string-arguments since there exists no way to
specify user-defined variables and therefore such demands would be
unnecessary.

Additionally this patch updates the existing tests (parse-line and
load-config) to test such cases where a string literal is specified as
an argument for a command in the config.